调整服务端代码以及逻辑并解决API工程编译问题（待测试）

+++ skywalking-collector/skywalking-protocol/src/main/java/com/ai/cloud/skywalking/util/IntegerAssist.java
@@ -15,10 +15,7 @@ public class IntegerAssist {
 
     public static int bytesToInt(byte[] src, int offset) {
         int value;
-        value = (int) (((src[offset] & 0xFF) << 24)
-                | ((src[offset + 1] & 0xFF) << 16)
-                | ((src[offset + 2] & 0xFF) << 8)
-                | (src[offset + 3] & 0xFF));
+        value = (((src[offset] & 0xFF) << 24) | ((src[offset + 1] & 0xFF) << 16) | ((src[offset + 2] & 0xFF) << 8) | (src[offset + 3] & 0xFF));
         return value;
     }
 }
